# src/docs/swagger.yaml
basePath: /
definitions:
  github_com_gothi_vouchrs_src_delivery_http_response.APIError:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: listing not found
        type: string
    type: object
  github_com_gothi_vouchrs_src_delivery_http_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.APIError'
      meta: {}
      success:
        example: true
        type: boolean
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequest:
    properties:
      alertedCount:
        type: integer
      brandID:
        type: string
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      maxPrice:
        type: number
      maxValue:
        type: number
      minValue:
        type: number
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequestStatus'
      userID:
        type: string
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequestStatus:
    enum:
    - active
    - fulfilled
    - expired
    - cancelled
    type: string
    x-enum-varnames:
    - BuyRequestStatusActive
    - BuyRequestStatusFulfilled
    - BuyRequestStatusExpired
    - BuyRequestStatusCancelled
  github_com_gothi_vouchrs_src_internal_domain_entity.CardRequest:
    properties:
      adminNotes:
        type: string
      brand:
        description: free-text brand name
        type: string
      createdAt:
        type: string
      desiredValue:
        type: number
      fulfilledAt:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestStatus'
      updatedAt:
        type: string
      urgency:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestUrgency'
      userID:
        type: string
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestStatus:
    enum:
    - PENDING_ADMIN_REVIEW
    - UNDER_REVIEW
    - FULFILLED
    - REJECTED
    - DEFERRED
    type: string
    x-enum-varnames:
    - CardRequestStatusPendingReview
    - CardRequestStatusUnderReview
    - CardRequestStatusFulfilled
    - CardRequestStatusRejected
    - CardRequestStatusDeferred
  github_com_gothi_vouchrs_src_internal_domain_entity.CardRequestUrgency:
    enum:
    - flexible
    - within_24h
    - within_1_week
    type: string
    x-enum-varnames:
    - CardRequestUrgencyFlexible
    - CardRequestUrgency24h
    - CardRequestUrgency1Week
...
  github_com_gothi_vouchrs_src_internal_domain_entity.Listing:
    properties:
      brandID:
        type: string
      buyerPrice:
        description: what buyer pays
        type: number
      codeEncrypted:
        description: AES-256-GCM encrypted
        type: string
      codeHash:
        description: SHA-256 for duplicate check
        type: string
      createdAt:
        type: string
      discountPct:
        description: e.g. 9.0 for 9% off
        type: number
      faceValue:
        description: e.g. 1000.00
        type: number
      gate1At:
        type: string
      id:
        type: string
      isPool:
        description: true = CardSwap pool
        type: boolean
      lockBuyerID:
        type: string
      lockExpiresAt:
        type: string
      sellerID:
        type: string
      sellerPayout:
        description: what seller receives
        type: number
      soldAt:
        type: string
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.ListingStatus'
      updatedAt:
        type: string
      verifiedBalance:
        description: balance at Gate 1
        type: number
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.ListingStatus:
    enum:
    - LIVE
    - LOCKED
    - SOLD
    - EXPIRED
    - CANCELLED
    - FRAUD_HOLD
    type: string
    x-enum-varnames:
    - ListingStatusLive
    - ListingStatusLocked
    - ListingStatusSold
    - ListingStatusExpired
    - ListingStatusCancelled
    - ListingStatusFraudHold
...
  github_com_gothi_vouchrs_src_internal_domain_entity.Transaction:
    properties:
      buyerAmount:
        type: number
      buyerID:
        type: string
      codeRevealedAt:
        description: when email was sent
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      id:
        type: string
      listingID:
        type: string
      lockStartedAt:
        type: string
      paidAt:
        type: string
      paymentRef:
        description: PG merchant transaction ID
        type: string
      payoutRef:
        description: Razorpay payout ID
        type: string
      sellerID:
        type: string
      sellerPayout:
        type: number
      status:
        $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.TransactionStatus'
      updatedAt:
        type: string
    type: object
  github_com_gothi_vouchrs_src_internal_domain_entity.TransactionStatus:
    enum:
    - pending
    - paid
    - completed
    - cancelled
    - refunded
    type: string
    x-enum-comments:
      TxnStatusCompleted: buyer confirmed redemption
      TxnStatusPaid: payment confirmed, code sent by email
      TxnStatusPending: lock acquired, waiting for payment
    x-enum-descriptions:
    - lock acquired, waiting for payment
    - payment confirmed, code sent by email
    - buyer confirmed redemption
    - ""
    - ""
    x-enum-varnames:
    - TxnStatusPending
    - TxnStatusPaid
    - TxnStatusCompleted
    - TxnStatusCancelled
    - TxnStatusRefunded
...
  github_com_gothi_vouchrs_src_internal_domain_port.MyRequestsResult:
    properties:
      buyRequests:
        items:
          $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.BuyRequest'
        type: array
      cardRequests:
        items:
          $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.CardRequest'
        type: array
    type: object
  src_delivery_http_handler.authUserDoc:
    properties:
      email:
        example: user@example.com
        type: string
      full_name:
        example: Rahul Sharma
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_verified:
        example: true
        type: boolean
      phone:
        example: "+919999999999"
        type: string
      role:
        example: buyer
        type: string
    type: object
...
  src_delivery_http_handler.meResponse:
    properties:
      email:
        example: user@example.com
        type: string
      role:
        example: buyer
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  src_delivery_http_handler.refreshTokenBody:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  src_delivery_http_handler.requestOTPBody:
    properties:
      contact:
        example: "+919999999999"
        type: string
    required:
    - contact
    type: object
...
  src_delivery_http_handler.tokenPairDoc:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  src_delivery_http_handler.verifyOTPBody:
    properties:
      contact:
        example: "+919999999999"
        type: string
      otp:
        example: "123456"
        type: string
    required:
    - contact
    - otp
    type: object
  src_delivery_http_handler.verifyOTPResponse:
    properties:
      tokens:
        $ref: '#/definitions/src_delivery_http_handler.tokenPairDoc'
      user:
        $ref: '#/definitions/src_delivery_http_handler.authUserDoc'
    type: object
host: localhost:8080
info:
  contact:
    email: support@cardswap.in
    name: CardSwap Support
  description: P2P gift card exchange marketplace. Card codes are AES-256 encrypted
    at rest and **never returned in API responses** — they are delivered to the buyer's
    registered email only.
  title: CardSwap India API
  version: "1.0"
paths:
  /api/v1/admin/auth/callback:
    get:
      description: Google redirects here after consent. Returns JWT tokens if the
        email is in the admin allowlist.
      parameters:
      - description: OAuth authorization code from Google
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tokens and admin user
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/src_delivery_http_handler.verifyOTPResponse'
              type: object
        "400":
          description: Missing code
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: Email not in admin allowlist
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: Admin — Google OAuth callback
      tags:
      - admin-auth
  /api/v1/admin/auth/login:
    get:
      description: Redirects the admin browser to Google's OAuth consent page. Open
        this URL directly in a browser.
      parameters:
      - description: CSRF state token
        example: '"cardswap_admin"'
        in: query
        name: state
        type: string
      responses:
        "307":
          description: Redirect to Google OAuth consent page
      summary: Admin — initiate Google OAuth login
      tags:
      - admin-auth
...
  /api/v1/auth/logout:
    post:
      description: Revoke the current access token. The token is invalidated immediately
        server-side.
      produces:
      - application/json
      responses:
        "200":
          description: logged out
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access + refresh token
        pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/src_delivery_http_handler.refreshTokenBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/src_delivery_http_handler.tokenPairDoc'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: Refresh token invalid or expired
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/request-otp:
    post:
      consumes:
      - application/json
      description: Send a 6-digit OTP to a phone number (SMS) or email address. Rate-limited
        to 5 attempts per 15 minutes.
      parameters:
      - description: Phone number or email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/src_delivery_http_handler.requestOTPBody'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "429":
          description: Too many attempts
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: Request OTP
      tags:
      - auth
  /api/v1/auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Validate the 6-digit OTP. Creates a new user on first login. Returns
        access + refresh JWT pair.
      parameters:
      - description: Contact and OTP
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/src_delivery_http_handler.verifyOTPBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/src_delivery_http_handler.verifyOTPResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "401":
          description: OTP invalid or expired
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
        "403":
          description: User is banned
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      summary: Verify OTP and receive JWT tokens
      tags:
      - auth
...
  /api/v1/dashboard/listings:
    get:
      description: Returns a paginated list of listings created by the authenticated
        seller. Sensitive code fields are stripped.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.Listing'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: My listings (seller dashboard)
      tags:
      - dashboard
  /api/v1/dashboard/purchases:
    get:
      description: Returns a paginated list of transactions where the authenticated
        user is the buyer.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_entity.Transaction'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: My purchases (buyer dashboard)
      tags:
      - dashboard
  /api/v1/dashboard/requests:
    get:
      description: Returns all buy requests and card requests created by the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gothi_vouchrs_src_internal_domain_port.MyRequestsResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: My requests (dashboard)
      tags:
      - dashboard
...
  /api/v1/users/me:
    get:
      description: Returns the authenticated user's ID, role, and email decoded from
        the JWT claims.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/src_delivery_http_handler.meResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gothi_vouchrs_src_delivery_http_response.Response'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - auth
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"